requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
sqlalchemy[asyncio]==2.0.36
asyncmy==0.2.9
alembic==1.13.3
pydantic-settings==2.3.4
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
httpx==0.27.2
itsdangerous==2.2.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
black==24.10.0
isort==5.13.2
ruff==0.6.9

# Optional: Redis for distributed session storage
# redis==5.0.1
# aioredis==2.0.1

